+++ Tags/Blamtoozle/halo3_tools_pc64/halo3_tools_pc64_tag_struct_definition.h
@@ -28,7 +28,7 @@ enum e_halo3_tools_pc64_tag_memory_usage_bit
 //	return (this->runtime_flags >> 6) & 1;
 //}
 
-enum e_halo3_tools_pc64_tag_field_set_bit : unsigned long
+enum e_halo3_tools_pc64_tag_field_set_bit : unsigned int
 {
 	_halo3_tools_pc64_tag_field_set_unknown0,
 	_halo3_tools_pc64_tag_field_set_unknown1,
