+++ Tags/Blamtoozle/halo4_x360/halo4_x360_tag_reference_definition.h
@@ -13,7 +13,7 @@ enum e_halo4_x360_tag_reference_definition_flags // #TODO: Check these. Strings 
 
 struct s_halo4_x360_tag_reference_definition
 {
-	c_flags<e_halo4_x360_tag_reference_definition_flags, unsigned long> flags;
+	c_flags<e_halo4_x360_tag_reference_definition_flags, unsigned int> flags;
 	tag group_tag;
 	ptr32 group_tags_address;
 };
