+++ Tags/Blamtoozle/reach_x360/reach_x360_tag_reference_definition.h
@@ -13,7 +13,7 @@ enum e_reach_x360_tag_reference_definition_flags // #TODO: Check these. Strings 
 
 struct s_reach_x360_tag_reference_definition
 {
-	c_flags<e_reach_x360_tag_reference_definition_flags, unsigned long> flags;
+	c_flags<e_reach_x360_tag_reference_definition_flags, unsigned int> flags;
 	tag group_tag;
 	ptr32 group_tags_address;
 };
